requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
mysql-connector-python==8.2.0
qrcode[pil]==7.4.2
Pillow>=10.2.0

